Refactor for microbenchmarks; builds and passes tests but has not been run
