+++ src/main/java/com/vmware/bespin/rpc/RPCServer.java
@@ -5,9 +5,9 @@
 
 package com.vmware.bespin.rpc;
 
-import com.vmware.bespin.scheduler.rpc.RPCID;
-
 import java.io.IOException;
+
+import com.vmware.bespin.scheduler.dinos.rpc.RPCID;
 
 /// RPC server operations
 public abstract class RPCServer<S> {
