+++ src/main/java/com/vmware/bespin/rpc/TCPClient.java
@@ -5,10 +5,11 @@
 
 package com.vmware.bespin.rpc;
 
-import com.vmware.bespin.scheduler.rpc.RPCID;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
- 
+
+import com.vmware.bespin.scheduler.dinos.rpc.RPCID;
+
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.IOException;
