+++ src/main/java/com/vmware/bespin/scheduler/SolverException.java
@@ -1,0 +1,8 @@
+package com.vmware.bespin.scheduler;
+
+public class SolverException 
+  extends Exception {
+    public SolverException(final String errorMessage, final Throwable err) {
+        super(errorMessage, err);
+    }
+}
