+++ src/main/java/com/vmware/bespin/scheduler/dinos/rpc/AffinityReleaseHandler.java
@@ -1,4 +1,4 @@
-package com.vmware.bespin.scheduler.rpc;
+package com.vmware.bespin.scheduler.dinos.rpc;
 
 import com.vmware.bespin.rpc.RPCHandler;
 import com.vmware.bespin.rpc.RPCHeader;
@@ -8,23 +8,23 @@ import com.vmware.bespin.rpc.RPCHeader;
  */
 
 import com.vmware.bespin.rpc.RPCMessage;
-import com.vmware.bespin.scheduler.DCMRunner;
+import com.vmware.bespin.scheduler.Scheduler;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 
-public class AffinityReleaseHandler extends RPCHandler<DCMRunner> {
+public class AffinityReleaseHandler extends RPCHandler<Scheduler> {
     private static final Logger LOG = LogManager.getLogger(AffinityReleaseHandler.class);
 
     public AffinityReleaseHandler() { }
 
     @Override
-    public RPCMessage handleRPC(final RPCMessage msg, final DCMRunner runner) {
+    public RPCMessage handleRPC(final RPCMessage msg, final Scheduler scheduler) {
         final RPCHeader hdr = msg.hdr();
         final AffinityRequest req = new AffinityRequest(msg.payload());
 
         // Add to the node capacity
-        runner.updateNode(req.nodeId, req.cores, req.memslices, true);
+        scheduler.updateNode(req.nodeId, req.cores, req.memslices, true);
 
         LOG.info("Processed scheduler affinity release request: {}", req); 
         hdr.msgLen = AffinityResponse.BYTE_LEN;
