+++ src/main/java/com/vmware/bespin/scheduler/dinos/rpc/AffinityResponse.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: BSD-2 OR MIT
  */
 
- package com.vmware.bespin.scheduler.rpc;
+ package com.vmware.bespin.scheduler.dinos.rpc;
  
  public class AffinityResponse {
      public static final int BYTE_LEN = 1;
