+++ src/main/java/com/vmware/bespin/scheduler/dinos/rpc/ReleaseHandler.java
@@ -3,26 +3,26 @@
  * SPDX-License-Identifier: BSD-2 OR MIT
  */
 
-package com.vmware.bespin.scheduler.rpc;
+package com.vmware.bespin.scheduler.dinos.rpc;
 
 import com.vmware.bespin.rpc.RPCHandler;
 import com.vmware.bespin.rpc.RPCHeader;
 import com.vmware.bespin.rpc.RPCMessage;
-import com.vmware.bespin.scheduler.DCMRunner;
+import com.vmware.bespin.scheduler.Scheduler;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class ReleaseHandler extends RPCHandler<DCMRunner> {
+public class ReleaseHandler extends RPCHandler<Scheduler> {
     private static final Logger LOG = LogManager.getLogger(ReleaseHandler.class);
 
     public ReleaseHandler() { }
 
     @Override
-    public RPCMessage handleRPC(final RPCMessage msg, final DCMRunner runner) {
+    public RPCMessage handleRPC(final RPCMessage msg, final Scheduler scheduler) {
         final RPCHeader hdr = msg.hdr();
         final ReleaseRequest req = new ReleaseRequest(msg.payload());
 
-        runner.releaseAllocation(req.nodeId, req.application, req.cores, req.memslices);
+        scheduler.releaseAllocation(req.nodeId, req.application, req.cores, req.memslices);
         LOG.info("Processed release request: {}", req);
 
         hdr.msgLen = ReleaseResponse.BYTE_LEN;
